+++ include/LIEF/DWARF/CompilationUnit.hpp
@@ -228,6 +228,25 @@ class LIEF_API CompilationUnit {
   /// `main`.
   functions_it functions() const;
 
+  /// Return an iterator over the functions **imported** in this compilation
+  /// unit **but not** implemented.
+  ///
+  /// For instance with this code:
+  ///
+  /// ```cpp
+  /// #include <cstdio>
+  /// int main() {
+  ///   printf("Hello\n");
+  ///   return 0;
+  /// }
+  /// ```
+  ///
+  /// `printf` is imported from the standard libc so the function is returned by
+  /// the iterator. On the other hand, `main()` is implemented in this
+  /// compilation unit so it is not returned by imported_function() but
+  /// functions().
+  functions_it imported_functions() const;
+
   /// Return an iterator over the different types defined in this
   /// compilation unit.
   types_it types() const;